+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 
 * Can tweak the number of cpu cores seen by the game
+* Vulkan: Use a fence to wait for each image to be presented (#741)
 
 ### Changed